Create swift CAM OSM MKOverlay
This code defines a CleanAppMapOverlay class that implements the MKOverlay protocol. The MKOverlay protocol requires conforming types to have a coordinate property and a boundingMapRect property, which define the location and size of the overlay on the map.

To use this CleanAppMapOverlay class, you would first create an instance of it and specify the coordinate and boundingMapRect properties, and then add it to an MKMapView using the addOverlay(_:) method. For example:

let cleanAppMapOverlay = CleanAppMapOverlay(coordinate: CLLocationCoordinate2D(latitude: 37.334722, longitude: -122.009167), boundingMapRect: MKMapRect(origin: MKMapPoint(x: 0, y: 0), size: MKMapSize(width: 10000, height: 10000)))
mapView.addOverlay(cleanAppMapOverlay)

This code creates an instance of CleanAppMapOverlay with a coordinate of (37.334722, -122.009167) and a boundingMapRect with an origin of (0, 0) and a size of (10000, 10000), and then adds it to an MKMapView.